Final Commit, Random CPUs
Abandoned the idea of using a Monte Carlo algorithm to determine the best possible move and instead started programming strategy into the Random CPUs.  RandomCPU still plays randomly, and RandomPlusCPU still blocks two-in-a-rows and will complete two-in-a-rows.  RandomPlusPlusCPU looks for places where playing would give them two two-in-a-rows and blocks places where playing would give an opponent two two-in-a-rows.  RandomPlusThreeCPU goes one step further by looking for places where playing would force the opponent to give us two two-in-a-rows or places where we might be able to get two two-in-a-rows and also by blocking places where an opponent could do this to them.  RandomPlusThreeCPU actually looks far enough ahead that when it plays against itself, the game always ends in a cat's game.  As such my work on this app is finished; my AI can never be beaten, even by itself.